Re-activate all RTS unit tests (except BigInt)

+++ rts/motoko-rts-tests/src/gc.rs
@@ -93,7 +93,7 @@ fn test_heaps() -> Vec<TestHeap> {
     ]
 }
 
-fn test_random_heap(seed: u64, max_objects: u32) {
+fn test_random_heap(seed: u64, max_objects: usize) {
     let random_heap = random::generate(seed, max_objects);
     test_gcs(&random_heap);
 }
@@ -175,7 +175,7 @@ fn initialize_gc(heap: &mut MotokoHeap) {
 
         // Synchronize the partitioned heap with one big combined allocation by starting from the base pointer as the heap pointer.
         let result =
-            get_partitioned_heap().allocate(heap, Bytes(allocation_size as u32).to_words());
+            get_partitioned_heap().allocate(heap, Bytes(allocation_size).to_words());
         // Check that the heap pointer (here equals base pointer) is unchanged, i.e. no partition switch has happened.
         // This is a restriction in the unit test where `MotokoHeap` only supports contiguous bump allocation during initialization.
         assert_eq!(result.get_ptr(), heap.heap_base_address());
@@ -190,7 +190,7 @@ fn reset_gc() {
     use crate::memory::TestMemory;
     use motoko_rts::gc::incremental::{partitioned_heap::PARTITION_SIZE, IncrementalGC};
 
-    let mut memory = TestMemory::new(Words(PARTITION_SIZE as u32));
+    let mut memory = TestMemory::new(Words(PARTITION_SIZE));
     unsafe {
         IncrementalGC::initialize(&mut memory, 0);
     }
@@ -228,17 +228,17 @@ fn check_dynamic_heap(
     let mut seen: FxHashMap<ObjectIdx, usize> = Default::default();
 
     let continuation_table_addr = unskew_pointer(read_word(heap, continuation_table_ptr_offset));
-    let continuation_table_offset = continuation_table_addr as usize - heap.as_ptr() as usize;
+    let continuation_table_offset = continuation_table_addr - heap.as_ptr() as usize;
 
     while offset < heap_ptr_offset {
         let object_offset = offset;
 
         // Address of the current object. Used for debugging.
-        let address = offset as usize + heap.as_ptr() as usize;
+        let address = offset + heap.as_ptr() as usize;
 
         if object_offset == continuation_table_offset {
             check_continuation_table(object_offset, continuation_table, heap);
-            offset += (size_of::<Array>() + Words(continuation_table.len() as u32))
+            offset += (size_of::<Array>() + Words(continuation_table.len()))
                 .to_bytes()
                 .as_usize();
             continue;
@@ -251,7 +251,7 @@ fn check_dynamic_heap(
             assert!(incremental);
         } else if tag == TAG_FREE_SPACE {
             assert!(incremental);
-            let words = read_word(heap, offset) as usize;
+            let words = read_word(heap, offset);
             offset += WORD_SIZE;
             offset += words * WORD_SIZE;
         } else {
@@ -260,16 +260,16 @@ fn check_dynamic_heap(
                 forward = read_word(heap, offset);
                 offset += WORD_SIZE;
             } else {
-                forward = make_pointer(address as u32);
+                forward = make_pointer(address);
             }
 
-            let is_forwarded = forward != make_pointer(address as u32);
+            let is_forwarded = forward != make_pointer(address);
 
             if incremental && tag == TAG_BLOB {
                 assert!(!is_forwarded);
                 // in-heap mark stack blobs
                 let length = read_word(heap, offset);
-                offset += WORD_SIZE + length as usize;
+                offset += WORD_SIZE + length;
             } else {
                 if incremental {
                     assert!(tag == TAG_ARRAY || tag >= TAG_ARRAY_SLICE_MIN);
@@ -280,14 +280,14 @@ fn check_dynamic_heap(
                 if is_forwarded {
                     assert!(incremental);
 
-                    let forward_offset = forward as usize - heap.as_ptr() as usize;
+                    let forward_offset = forward - heap.as_ptr() as usize;
                     let length = read_word(
                         heap,
                         forward_offset + size_of::<Obj>().to_bytes().as_usize(),
                     );
 
                     // Skip stale object version that has been relocated during incremental GC.
-                    offset += length as usize * WORD_SIZE;
+                    offset += length * WORD_SIZE;
                 } else {
                     let n_fields = read_word(heap, offset);
                     offset += WORD_SIZE;
@@ -316,12 +316,12 @@ fn check_dynamic_heap(
                         offset += WORD_SIZE;
                         // Get index of the object pointed by the field
                         let pointee_address = field.wrapping_add(1); // unskew
-                        let pointee_offset = (pointee_address as usize) - (heap.as_ptr() as usize);
+                        let pointee_offset = pointee_address - heap.as_ptr() as usize;
                         let pointee_idx_offset =
-                            pointee_offset as usize + size_of::<Array>().to_bytes().as_usize(); // skip array header (incl. length)
+                            pointee_offset + size_of::<Array>().to_bytes().as_usize(); // skip array header (incl. length)
                         let pointee_idx = get_scalar_value(read_word(heap, pointee_idx_offset));
                         let expected_pointee_idx =
-                            object_expected_pointees[(field_idx - 1) as usize];
+                            object_expected_pointees[field_idx - 1];
                         assert_eq!(
                             pointee_idx,
                             expected_pointee_idx,
@@ -422,11 +422,11 @@ fn check_continuation_table(mut offset: usize, continuation_table: &[ObjectIdx],
     offset += WORD_SIZE;
 
     if incremental {
-        assert_eq!(read_word(heap, offset), make_pointer(table_addr as u32));
+        assert_eq!(read_word(heap, offset), make_pointer(table_addr));
         offset += WORD_SIZE;
     }
 
-    assert_eq!(read_word(heap, offset), continuation_table.len() as u32);
+    assert_eq!(read_word(heap, offset), continuation_table.len());
     offset += WORD_SIZE;
 
     for obj in continuation_table.iter() {
@@ -434,7 +434,7 @@ fn check_continuation_table(mut offset: usize, continuation_table: &[ObjectIdx],
         offset += WORD_SIZE;
 
         // Skip object header for idx
-        let idx_address = ptr as usize + size_of::<Array>().to_bytes().as_usize();
+        let idx_address = ptr + size_of::<Array>().to_bytes().as_usize();
         let idx = get_scalar_value(read_word(heap, idx_address - heap.as_ptr() as usize));
 
         assert_eq!(idx, *obj);
@@ -444,7 +444,7 @@ fn check_continuation_table(mut offset: usize, continuation_table: &[ObjectIdx],
 impl GC {
     #[non_incremental_gc]
     fn run(&self, heap: &mut MotokoHeap, _round: usize) -> bool {
-        let heap_base = heap.heap_base_address() as u32;
+        let heap_base = heap.heap_base_address();
         let static_roots = Value::from_ptr(heap.static_root_array_address());
         let continuation_table_ptr_address = heap.continuation_table_ptr_address() as *mut Value;
 
@@ -460,7 +460,7 @@ impl GC {
                         // get_hp
                         || heap_1.heap_ptr_address(),
                         // set_hp
-                        move |hp| heap_2.set_heap_ptr_address(hp as usize),
+                        move |hp| heap_2.set_heap_ptr_address(hp),
                         static_roots,
                         continuation_table_ptr_address,
                         // note_live_size
@@ -480,7 +480,7 @@ impl GC {
                         // get_hp
                         || heap_1.heap_ptr_address(),
                         // set_hp
-                        move |hp| heap_2.set_heap_ptr_address(hp as usize),
+                        move |hp| heap_2.set_heap_ptr_address(hp),
                         static_roots,
                         continuation_table_ptr_address,
                         // note_live_size
@@ -505,10 +505,10 @@ impl GC {
                 };
                 unsafe {
                     REMEMBERED_SET = Some(RememberedSet::new(heap));
-                    LAST_HP = heap_1.last_ptr_address() as u32;
+                    LAST_HP = heap_1.last_ptr_address();
 
                     let limits = motoko_rts::gc::generational::Limits {
-                        base: heap_base as usize,
+                        base: heap_base,
                         last_free: heap_1.last_ptr_address(),
                         free: heap_1.heap_ptr_address(),
                     };
